beernary/requirements.txt
RPi.GPIO
Mock.GPIO
pySerial
ConfigParser
loguru
PyMySQL